aprgcommon refactoring

+++ AprgCommon/AprgCommon/tst/CommonTests/Randomizer/AlbaUniformDeterministicRandomizer_unit.cpp
@@ -46,15 +46,14 @@ TEST(AlbaUniformDeterministicRandomizerTest, GetRandomIntegerWorks_AsUniformlyDi
     int numberOfRandomValues(maximumValue - minimumValue + 1);
     int const iterations(1000);
     int const allowedDeviationCount(iterations / 10);
-    vector<int> hitsForEachValue(static_cast<unsigned long long>(numberOfRandomValues), 0);
+    vector<int> hitsForEachValue(numberOfRandomValues, 0);
 
     for (int i = 0; i < iterations; i++) {
         auto randomValue(randomizer.getRandomValue());
-        hitsForEachValue[static_cast<unsigned long long>(randomValue)]++;
+        hitsForEachValue[randomValue]++;
     }
 
-    ASSERT_FALSE(hitsForEachValue.empty());
-    auto&& [minIterator, maxIterator] = std::minmax_element(hitsForEachValue.cbegin(), hitsForEachValue.cend());
+    ASSERT_FALSE(hitsForEachValue.empty());    auto&& [minIterator, maxIterator] = std::minmax_element(hitsForEachValue.cbegin(), hitsForEachValue.cend());
     int deviationCount(*maxIterator - *minIterator);
     EXPECT_LT(deviationCount, allowedDeviationCount);
 }
@@ -97,11 +96,10 @@ TEST(AlbaUniformDeterministicRandomizerTest, GetRandomFloatingValueWorks_AsUnifo
 
     for (int i = 0; i < iterations; i++) {
         auto randomValue(randomizer.getRandomValue());
-        hitsForEachValue[static_cast<unsigned long long>(randomValue)]++;
+        hitsForEachValue[randomValue]++;
     }
 
-    ASSERT_FALSE(hitsForEachValue.empty());
-    auto&& [minIterator, maxIterator] = std::minmax_element(hitsForEachValue.cbegin(), hitsForEachValue.cend());
+    ASSERT_FALSE(hitsForEachValue.empty());    auto&& [minIterator, maxIterator] = std::minmax_element(hitsForEachValue.cbegin(), hitsForEachValue.cend());
     int deviationCount(*maxIterator - *minIterator);
     EXPECT_LT(deviationCount, allowedDeviationCount);
 }
